Add files via upload
In this project, we analyzed toxic comments from the internet using NLP models to build a classification model. We utilized data provided by Wikipedia to read and merge training and testing datasets. The text was cleaned, and a word cloud was created for visualization. After vectorizing the text, we defined the target and features, training our model with the training data and testing it with the test dataset. The model performed successfully, and we visualized the results using a confusion matrix heatmap to assess its performance.